==> notebooks/00.exploration.ipynb <==
DATA_PATH = 'data/raw/raw_data.csv'

import numpy as np
import pandas as pd
pd.options.display.max_rows = 10000

df = pd.read_csv(DATA_PATH)
df.head()

df.shape

df.info()

df.describe()

df.sample(1).iloc[0]

questionable_columns = ['YearsCode', 'YearsCodePro']
for col_name in questionable_columns:
    col = df[col_name]
    print(f' "{col_name}" unique values: \n {col.unique().tolist()} \n')

* the fields 'More than 50 years' and 'Less than 1 year' in the above columns need to be handeled
* fields that are separated by delimiters need to be handled
