==> notebooks/01.preprocessing.ipynb <==
DATA_PATH = '../data/raw/raw_data.csv'
EXPORT_PATH= '../data/processed/preprocessed_data.pkl'

import numpy as np
import pandas as pd
pd.options.display.max_rows = 10000
from sklearn.preprocessing import MultiLabelBinarizer
import pickle

## Functions

def split_answers(data_series, delimiters=[';', ',']):
    """
    split multiple answers in a string to a list based on delimiters - only one delimiter is evaluated at most
    which is the first match in the list
    parameters:
        data_series(pd.Series):string series with some answers
        delimiters(list): list of strings representing the delimiters used in the split - default =[';' ',']
    Returns:
        (pd.Series) : a series containing the splitted data if applicable other wise the original series
    """
    
    def check_splittable(series, delimiters):
        for delimiter in delimiters:
            if series.str.contains(delimiter).any():
                return (True, delimiter)
        return (False, None)
    
    def split(series, delimiter):
        return series.str.split(delimiter)
    
    is_splittable, delimiter = check_splittable(data_series, delimiters)
    if not is_splittable:
        return data_series
    
    splitted_series = split(data_series, delimiter)
    
    # replace na with empty list
    na_mask = splitted_series.isnull()
    splitted_series.loc[na_mask] = splitted_series.loc[na_mask].apply(lambda x:[])
    
    return splitted_series

REPLACE_DICT = {
    'YearsCodePro': {'Less than 1 year': 0, 'More than 50 years': 51}, 
    'YearsCode':    {'Less than 1 year': 0, 'More than 50 years': 51}}

## Preprocessing

raw_df = pd.read_csv(DATA_PATH)
df = raw_df.copy()

#### replacing values

for col, replacments in REPLACE_DICT.items():
    df[col] = df[col].replace(replacments).astype(np.float32)

#### splitting multiple answers

obj_cols = df.select_dtypes(include='object').columns.tolist()
for col in obj_cols:
    df[col] = split_answers(df[col])

## sanity check

df.sample(1).iloc[0]

## Exporting data

df.to_pickle(EXPORT_PATH)
