==> notebooks/04a.model_baseline.ipynb <==
DF_PATH       = '../data/processed/merged_cleaned_data.pkl'

ROLE_COLS      = ['DevType']
TECH_COLS      = ['LanguageHaveWorkedWith',
                  'DatabaseHaveWorkedWith',
                  'PlatformHaveWorkedWith',
                  'WebframeHaveWorkedWith',
                  'MiscTechHaveWorkedWith',
                  'ToolsTechHaveWorkedWith',
                  'NEWCollabToolsHaveWorkedWith']

MLFLOW_TRACKING_URI = '../models/mlruns'
MLFLOW_EXPERIMENT_NAME = "skills_jobs_stackoverflow"

LOG_PATH = "../models/temp/"
LOG_DATA_PKL    =  "data.pkl"
LOG_MODEL_PKL   =  "model.pkl"
LOG_METRICS_PKL =  "metrics.pkl"

# Load packages
import pandas as pd 
import numpy as np
import logging
import pickle
import random
import os
import warnings
from pathlib import Path

import mlflow
from mlflow.tracking import MlflowClient

from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import make_pipeline
from sklearn.multioutput import MultiOutputClassifier
from sklearn.model_selection import train_test_split

from sklearn import metrics
from sklearn.metrics import auc, accuracy_score, confusion_matrix, f1_score, precision_score, recall_score

from sklearn.linear_model import LogisticRegression

warnings.filterwarnings("ignore")

## Functions

def calculate_quality(targets, predictions, metric_funciton, sort_values=False):
    """
    given target, predictions, and metric function it applies the metric
    to each column of the data and return the result as a data series
    Params:
        targets(pd.DataFrame): targets
        predictions(pd.DataFrame): predicions
        metric_function(sklearn.metrics): metric function
        sort_values(bool): if True the returned values will be sorted - default=False
    Returns:
        (pd.Series):metric score for each column
    """
    quality_scores={}
    for col in targets.columns:
        col_target = targets[col].copy()
        col_pred = predictions[col].copy()
        quality_scores[col] = round(metric_funciton(col_target, col_pred) *100, 2)
    
    quality_scores = pd.Series(quality_scores.values(), index=quality_scores.keys())
    
    if sort_values:
        quality_scores = quality_scores.sort_values()
    
    return quality_scores

def balance_sample(targets_df, samples_per_class=1200):
    """
    given a targets data frame returns a mask for balanced sample from the target classes
    Params:
        targets_df(pd.DataFrame): a one-hot encoded data frame with each column represent class
        samples_per_class(int): number of sampler per class - default=1200
    
    Returns:
        (pd.Series): sampling mask
    """
    resampled_targets = []
    for col in targets_df.columns:
        sub_df = targets_df.loc[targets_df[col]==1].copy()
        
        #upsampling
        if len(sub_df) < samples_per_class:
            sub_df = sub_df.sample(samples_per_class, replace=True, random_state=0)
        #downsampling
        else:
            sub_df = sub_df.sample(samples_per_class, random_state=0)
        resampled_targets.append(sub_df)
        
    resampled_targets = pd.concat(resampled_targets)
    
    return resampled_targets

def evaluate_model(model, X_train, X_test, Y_train, Y_test, metrics=[accuracy_score, precision_score, recall_score, f1_score]):
    """
    evaluates the model on the train and test data using the given metrics and report the results
    Params:
        model(sklearn.base)
        X_trian(np.ndarray)
        X_test(np.ndarray)
        Y_train(np.ndarray)
        Y_test(np.ndarray)
        metrics(list[sklearn.metrics]): default = [accuracy_score, precision_score, recall_score, f1_score]
    Returns:
        train_scores(pd.DataFrame)
        test_scores(pd.DataFrame)
    """
    
    # evaluate on training data
    predictions = pd.DataFrame( model.predict(X_train.values), columns=Y_train.columns)
    train_scores = {score.__name__: calculate_quality(Y_train, predictions, score) 
                    for score in metrics}
    train_scores = pd.concat(train_scores, axis=1)
    
    # evaluate on test data
    predictions = pd.DataFrame( model.predict(X_test.values), columns=Y_test.columns)
    test_scores = {score.__name__: calculate_quality(Y_test, predictions, score) 
                    for score in metrics}
    test_scores = pd.concat(test_scores, axis=1)
    
    return train_scores, test_scores

## Baseline Model

df = pd.read_pickle(DF_PATH)
df.head()

### Handling imbalanced classes

roles_df = df['DevType'].copy()
roles_df = balance_sample(roles_df)
df = df.loc[roles_df.index].copy()

### split the data

X_train, X_test, Y_train, Y_test = train_test_split(df.drop("DevType", axis=1), df["DevType"], random_state=0)

## Training

#### initializing MLFlow

# create directories
Path(MLFLOW_TRACKING_URI).mkdir(parents=True, exist_ok=True)
Path(LOG_PATH).mkdir(parents=True, exist_ok=True)

#initialize
mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()
exp = client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)
if not exp:
    mlflow.create_experiment(MLFLOW_EXPERIMENT_NAME)
    exp = client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)

#### Logistic Regression

lr_clf = make_pipeline(StandardScaler(),
                       MultiOutputClassifier(LogisticRegression()))
lr_clf.fit(X_train.values,Y_train.values)

train_scores, test_scores = evaluate_model(lr_clf, X_train, X_test, Y_train, Y_test)

mean_test_scores = test_scores.mean(axis=0)
mean_test_scores

test_scores

## Log run

# Data details
data_details = {"data_path": DF_PATH,
                "training_indices": X_train.index.tolist(),
                "test_indices":     X_test.index.tolist(), 
                "features_names":   X_train.columns.droplevel(0).tolist(),
                "targets_names":    Y_train.columns.tolist()}

with open(os.path.join(LOG_PATH, LOG_DATA_PKL), "wb") as output_file:
    pickle.dump(data_details, output_file)

# Model
model = {"model_description": "Baseline model: Logistic Regression ",
         "model_details": str(lr_clf),
         "model_object": lr_clf} 

with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), "wb") as output_file:
    pickle.dump(model, output_file)

# Performance details
classes_metrics = {"train_scores": train_scores, 
                   "test_scores":  test_scores}

with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), "wb") as output_file:
    pickle.dump(classes_metrics, output_file)

with mlflow.start_run(experiment_id=exp.experiment_id, run_name=model['model_description']):
    mlflow.log_artifacts(LOG_PATH)
    for metric, score in mean_test_scores.items():
        mlflow.log_metric(metric, score)
