==> notebooks/05.prediction.ipynb <==
MLFLOW_TRACKING_URI = '../models/mlruns'
MLFLOW_RUN_ID = "7a24a21a88fe4329a986ba8dd6c942cb"

LOG_DATA_PKL    =  "data.pkl"
LOG_MODEL_PKL   =  "model.pkl"
LOG_METRICS_PKL =  "metrics.pkl"

CLUSTERS_YAML_PATH = "../data/processed/fe_cluster_skills_description.yaml"

import os 
import sklearn
import pickle
import yaml
import re
import pandas as pd

import mlflow
from mlflow.tracking import MlflowClient

## Initialization

### MLFlow

mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()

run = mlflow.get_run(MLFLOW_RUN_ID)
artificats_path = re.sub('file:\\\\','',os.path.normpath(run.info.artifact_uri))

artificats_path

### Loading Model and Data

# model
model_path = os.path.join(artificats_path, LOG_MODEL_PKL)
with open(model_path, "rb") as f:
    model = pickle.load(f)

model

# Load data pkl
data_path  = os.path.join(artificats_path, LOG_DATA_PKL)
with open(data_path, 'rb') as handle:
    data = pickle.load(handle)

data.keys()

# Unpack vars
features_names = pd.Series(data["features_names"])
targets_names  = pd.Series(data['targets_names'])
classifier = model['model_object']

classifier

## Loading skill clusters

with open(CLUSTERS_YAML_PATH, "r") as stream:
    clusters_config = yaml.safe_load(stream)

clusters_list = [(cluster_name, skill)
                 for cluster_name, cluster_skills in clusters_config.items()
                 for skill in cluster_skills]
clusters_df = pd.DataFrame(clusters_list, columns=['cluster_name', 'skill'])

### Prediction

sample_skills = ['Scala', 'Hadoop', 'Python']
#Verification
pd.Series(sample_skills).isin(features_names).all()

### Creating cluster features

sample_clusters = clusters_df.copy()
sample_clusters["sample_skills"] = sample_clusters["skill"].isin(sample_skills)
cluster_features  = sample_clusters.groupby('cluster_name')['sample_skills'].sum()
cluster_features 

### Creating one-hot encoded skills

skills_names = features_names[~features_names.isin(cluster_features.index)]
skills_names

ohe_skills = pd.Series(skills_names.isin(sample_skills).astype(int).tolist(), index=skills_names)

features = pd.concat([ohe_skills, cluster_features]).loc[features_names]

#### Predict

predictions = classifier.predict_proba([features.values])

positive_probs = [prob[0][1] for prob in predictions]
pd.Series(positive_probs, 
          index=targets_names).sort_values(ascending=False)[:5]
