==> notebooks/06.simulation.ipynb <==
import sys
import pandas as pd
import sys

sys.path.append('../')

from scripts.prediction import *

## Simulation

### Calculate base preidction

current_skills = ['Python', 'Pandas']

model = JobPrediction()

base_predictions = model.predict_jobs_probabilities(current_skills).sort_values(ascending=False)

### Calculate probs for new skills

all_skills = pd.Series(model.get_all_skills())
new_skills = all_skills[~all_skills.isin(current_skills)].copy()

simulated_results = []
for skill in new_skills:
    additional_skill_prob = model.predict_jobs_probabilities([skill] + current_skills)
    additional_skill_uplift = (additional_skill_prob - base_predictions) / base_predictions
    additional_skill_uplift.name = skill
    simulated_results.append(additional_skill_uplift)

simulated_results_df = pd.DataFrame(simulated_results)
simulated_results_df

target_job = 'Data scientist or machine learning specialist'
target_results = simulated_results_df[target_job].sort_values(ascending=False)
target_results
